RES-2456: Move 'archive' framework to standalone package
