+++ MANIFEST.in
@@ -21,4 +21,3 @@
 # ------------------------------------------------------------------------------
 include README.md
 include LICENSE
-include requirements.txt
