+++ packages/archive/nupic/research/archive/dynamic_sparse/common/utils.py
