+++ packages/archive/nupic/research/archive/dynamic_sparse/networks/utils.py
