+++ packages/archive/nupic/research/archive/dynamic_sparse_compat/models/stochastic_synapses.py
