+++ packages/archive/nupic/research/archive/dynamic_sparse_compat/networks/binary_cnn.py
