+++ packages/archive/nupic/research/archive/dynamic_sparse_compat/networks/hard_concrete_cnn.py
