+++ packages/archive/nupic/research/archive/experiments_platform/networks/__init__.py
@@ -17,3 +17,6 @@
 #
 #  http://numenta.org/licenses/
 #
+
+from nupic.research.archive.dynamic_sparse.networks import *
+from nupic.research.frameworks.stochastic_connections.dynamic_sparse_compat.networks import *
