+++ packages/archive/setup.cfg
@@ -1,6 +1,6 @@
 # ------------------------------------------------------------------------------
 #  Numenta Platform for Intelligent Computing (NuPIC)
-#  Copyright (C) 2019, Numenta, Inc.  Unless you have an agreement
+#  Copyright (C) 2021, Numenta, Inc.  Unless you have an agreement
 #  with Numenta, Inc., for a separate license for this software code, the
 #  following terms and conditions apply:
 #
@@ -19,4 +19,16 @@
 #  http://numenta.org/licenses/
 #
 # ------------------------------------------------------------------------------
-__version__ = "0.0.1.dev0"
+[metadata]
+name = nupic.research.archive
+version = 0.0.1.dev0
+
+[options]
+zip_safe = False
+packages = find_namespace:
+install_requires =
+    nupic.research
+
+[options.packages.find]
+include = nupic.research.*
+
