+++ projects/dynamic_sparse/environment.yml
@@ -1,0 +1,26 @@
+# This file may be used to create an environment using:
+# $ conda env create --name <env> --file <this file>
+# platform: osx-64
+name: nupic.research.dynamic_sparse
+channels:
+  - defaults
+  - pytorch
+  - nvidia
+  - conda-forge
+
+dependencies:
+  - python=3.8
+  - pip
+
+  # FIXME: remove numba installation command once dependency issues with numpy 1.21.2 is fixed on a new numba release
+  #        Caused by this PR https://github.com/numba/numba/pull/7209
+  - numba=0.53.0
+  - pip:
+      # Install nupic.torch and nupic.rearch libraries in develop mode.
+      # Assume libraries were cloned into ${HOME}/nta folder
+      - -e ${HOME}/nta/nupic.torch
+      - -e ${HOME}/nta/nupic.research[dev]
+
+      # Install projects extra dependency
+      - -e ${HOME}/nta/nupic.research/packages/archive
+
