+++ setup.cfg
@@ -21,7 +21,7 @@
 # ------------------------------------------------------------------------------
 [metadata]
 name = nupic.research
-version = attr: nupic.research.__version__
+version = 0.0.1.dev0
 url = https://github.com/numenta/nupic.research
 project_urls =
     Documentation = https://nupicresearch.readthedocs.io
@@ -60,23 +60,23 @@ install_requires =
     h5py
     filelock
     boto3
-
-[options.extras_require]
-# Project specific dependencies
-#
-vernon =
+    
+    # vernon
     ray[tune]~=0.8.3
     seaborn
     wandb
     sigopt
-    
+
+[options.extras_require]
+# Project specific dependencies
+#
 imagenet =
     pretrainedmodels==0.7.4
     ax-platform
     botorch
     gpytorch
     pytorch-lightning==1.2.9
-    nupic.research[vernon]
+    nupic.research
 
 gsc =
     librosa==0.7.1
@@ -85,7 +85,7 @@ dendrites =
     ptitprince==0.2.5
     seaborn==0.11.1
     sklearn
-    nupic.research[vernon]
+    nupic.research
 
 transformers =
     ray[tune]==1.4.0
@@ -157,7 +157,7 @@ dev =
 
 
 [options.packages.find]
-include = nupic.*
+include = nupic.research.*
 
 [bdist_wheel]
 universal=0
